refactor(host): move the MCP tools dialog to its own file
PluginDrawerItemView had grown to hold the drawer item, the MCP badge, the
tool browser, the call history, and their context menus. Move everything
from the dialog down into McpToolsDialog.kt so each file covers one screen
concern. Code is unchanged apart from the visibility needed across the file
boundary and the import split.
